refactor: 도메인 순수함수 모듈 분리 (#1)
