+++ src/pages/SavingsCalculatorPage/domain/index.ts
@@ -1,0 +1,1 @@
+export * from './savingsProducts';
