+++ src/pages/SavingsCalculatorPage/domain/savingsProducts.ts
