+++ src/pages/SavingsCalculatorPage/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './format';
-export * from './savings';
